+++ np.Templating/CHANGELOG.md
@@ -20,6 +20,8 @@ See Plugin [Documentation](https://noteplan.co/templates/docs) for details on av
 - Added `editSettings` command to allow for mobile editing of plugin settings
 - Fix long-standing bug where date.format did not work correctly
 - Fix templaterunner bug where the file was not opening in the Editor
+- Add <current> to templateAppend command for easy testing of templates
+- Add `journalingQuestion` commands to WebModule per Tim Shaker - https://discord.com/channels/763107030223290449/963950027946999828/1051665188648132648
 
 ## [1.12.0] 2025-03-09 @dwertheimer